// src/interfaces/token.interface.ts
/* This TypeScript code snippet is defining interfaces and types related to token management. Here is a
breakdown of what each part of the code is doing: */

import { Document, Model } from "mongoose";
import { JwtPayload } from "jsonwebtoken";

export interface IToken {
  token: string;
  user: string;
  type: string;
  expires: Date;
}

export type NewToken = Omit<IToken, "blacklisted">;

export interface ITokenDoc extends IToken, Document {}

export interface ITokenModel extends Model<ITokenDoc> {}

export interface IPayload extends JwtPayload {
  sub: string;
  iat: number;
  exp: number;
  type: string;
}

export interface TokenPayload {
  token: string;
  expires: Date;
}

export interface AccessAndRefreshTokens {
  access: TokenPayload;
  refresh: TokenPayload;
}

export const tokenTypes = {
  ACCESS: "access",
  REFRESH: "refresh",
};